+++ rollup.config.js
@@ -1,34 +1,49 @@
-import typescript from 'rollup-plugin-typescript2';
-import resolve from 'rollup-plugin-node-resolve';
-import commonjs from 'rollup-plugin-commonjs';
-import uglify from 'rollup-plugin-uglify';
-import filesize from 'rollup-plugin-filesize';
-import peerDeps from 'rollup-plugin-peer-deps-external';
+import typescript from 'rollup-plugin-typescript2'
+import resolve from 'rollup-plugin-node-resolve'
+import commonjs from 'rollup-plugin-commonjs'
+import peerDeps from 'rollup-plugin-peer-deps-external'
+import uglify from 'rollup-plugin-uglify'
+import filesize from 'rollup-plugin-filesize'
 
-const format = process.env.NODE_ENV;
-const isUmd = format === 'umd';
-const file = `dist/redux-zero.${isUmd ? 'min' : format}.js`
+const format = process.env.NODE_ENV
+const isUmd = format === 'umd'
 
-const config = {
-  input: './src/index.ts',
-  name: 'redux-zero',
-  sourcemap: true,
-  output: {
-    file,
-    format,
-  },
-  plugins: [
-    peerDeps(),
-    typescript(),
-    resolve({
-      jsnext: true,
-      main: true,
-      browser: true,
-    }),
-    commonjs(),
-  ],
-};
+function getFileName(file) {
+  if (isUmd) {
+    return `${file}.min.js`
+  }
+  return `${file}.js`
+}
 
-isUmd && config.plugins.push(uglify(), filesize());
+function getConfig(input, file) {
+  const conf = {
+    input,
+    name: 'redux-zero',
+    sourcemap: true,
+    output: {
+      file: getFileName(file),
+      format,
+    },
+    plugins: [
+      peerDeps(),
+      typescript(),
+      resolve({
+        jsnext: true,
+        main: true,
+        browser: true,
+      }),
+      commonjs(),
+    ],
+  }
 
-export default config;
+  isUmd && conf.plugins.push(uglify(), filesize())
+
+  return conf
+}
+
+const config = [
+  getConfig('./src/index.ts', 'dist/redux-zero'),
+  getConfig('./src/react/index.ts', 'react/index'),
+]
+
+export default config