+++ src/index.ts
@@ -1,5 +1,3 @@
 import createStore from "./store/createStore"
-import connect from "./components/connect"
-import Provider from "./components/Provider"
 
-export { createStore, connect, Provider }
+export default createStore