+++ src/react/components/Provider.tsx
@@ -1,8 +1,8 @@
 import * as React from "react"
 
-import Props from "../interfaces/Props"
-import Store from "../interfaces/Store"
-import propValidation from "../utils/propsValidation"
+import Props from "../../interfaces/Props"
+import Store from "../../interfaces/Store"
+import propValidation from "../../utils/propsValidation"
 
 export default class Provider extends React.Component<Props, {}> {
   static childContextTypes = {